Fixed totals on view daysheet page.

+++ database/migrations/2022_11_13_164132_create_clients_table.php
@@ -16,7 +16,7 @@ class CreateClientsTable extends Migration
         Schema::create('clients', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name')->unique();
-            $table->float('markup')->default(0);
+            $table->decimal('markup', 8, 2)->default(0);
             $table->timestamps();
         });
     }
