+++ database/migrations/2023_03_05_113433_create_daysheets_table.php
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->boolean('client_confirmed')->default(false);
             $table->dateTime('confirmed_on')->nullable();
             $table->string('client_representative')->nullable();
-            $table->float('mileage_rate');
-            $table->float('markup_rate');
+            $table->decimal('mileage_rate', 8, 2);
+            $table->decimal('markup_rate', 8, 2);
             $table->timestamps();
         });
     }
